Put the channel reporter in its own feature
This also makes the crossbeam-channel optional (required just for the channel reporter implementation). Without the channel_reporter features, this crate consists of just the primary traits and it's skeleton, without any ready to use implementation. The channel_reporter feature is enabled by default for ease of getting started.
