+++ Cargo.toml
@@ -14,12 +14,20 @@ msrv = "1.38"
 exclude = ["/.github", "docs/sketches/*.png"]
 
 [features]
+default = ["channel_reporter"]
+channel_reporter = ["crossbeam-channel"]
+
 experimental_handle_disconnect_ack = []
 
-[dependencies]
-crossbeam-channel = "0.5"
+[dependencies.crossbeam-channel]
+version = "0.5"
+optional = true
 
 [dev-dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 indicatif = "0.16.2"
+
+[[example]]
+name = "json_lines"
+required-features = ["channel_reporter"]
